Add files via upload
I made this game in my first semester from "INTRO TO INFORMATION & COMMUNICATION TECH " course that was my first mini-project based on C language i.e. "Snake-Water-Gun Game" that is one of the famous two-player game played by many people. It is a hand game in which the player randomly chooses any of the three forms i.e. snake, water, and gun. We have all play this in the childhood. This is same game but computer play with you. You can play once, u will definitely enjoy it. This game similar to rock paper scissor.

These are the rules need to be followed while playing this game.
1) In the case of Snake vs. Water, Snake drinks the water hence wins.
2) In the case of Water vs. Gun, The gun will drown in water, hence a point for water
3) In the case of Gun vs. Snake, Gun will kill the snake and win.